Make the upload location of avatars configurable
The upload location of avatars via the `FileUploadService` is now
configurable through the `UPLOADED_AVATARS_PATH` Django setting and the
`get_unique_filename` method. If the `UPLOADED_AVATARS_PATH` setting is
not set, a default of `uploaded/avatars` will be used.

To enable this customizability, `AvatarServiceConfigForm`s are now
passed the instance of the service that instantiated them as a keyword
argument. This allows the forms to look up static settings on the
services.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8841/