Capture and report errors when extension media installation fails.
It's possible (and not very hard) to get into situations where extension
media installation fails. For instance, if there are
permission/ownership issues, we can fail to remove existing extension
media and crash when installing new media. This results in HTTP 500
error pages as each access then tries to repeat the media installation,
unsuccessfully.

We now look for these failure cases, catching the errors and reporting
them in a more helpful way. The pages no longer crash (but may not work
as expected), and admins can get a hint as to what's wrong in the error
logs.

There's also a fix for an exception handler's error reporting where the
error attempted to use the exception instance, which wasn't defined
during the capture.

Testing Done:
Tested against a failure case on our server.

Reviewed at https://reviews.reviewboard.org/r/8926/