Fix a regression in clearing the app cache on extension initialization.
The work done to improve extension management, loading, and Django
compatibility in Djblets 0.10 regressed clearing the app cache state for
Django 1.6. We were only resetting the app cache's loaded flag if
attempting to update the database, which wasn't enough. Amongst other
possible regressions, this prevented evolutions from being able to be
updated.

The clearing has now been moved out of the database upgrade code and
into the methods for manipulating the list of installed apps.

Testing Done:
Verified in a shell that the extensions were listed as part of a
`get_apps()` call, which wasn't the case before.

Deployed this to reviews.reviewboard.org, and verified that it was
including our extensions when considering database upgrades.

Reviewed at https://reviews.reviewboard.org/r/8927/