Add an input widget with a copy text link
Sometimes it makes sense to have a text widget that the user will want
to easily copy data from, such as with generated authentication tokens,
etc. The `CopyableTextInput` is a specialized `TextInput` widget that
renders a link to copy the text from the associated field.

Testing Done:
Using this with an upcoming change for OAuth forms.

Reviewed at https://reviews.reviewboard.org/r/8980/