Allow less-imports.js to be used in installed Djblets builds.
The less-imports.js was being shipped in the contrib/ directory. While
fine for development builds and eggs, it didn't work at all for wheels.
To fix this, less-imports.js has been moved into the package alongside
the custom LessCSS compiler, and the path has been updated to point to
this.

Testing Done:
Tested this on the build server. Before this, the build was failing
since the script could not be found. After, it successfully found the
script and executed it.

Reviewed at https://reviews.reviewboard.org/r/9011/