Fix building extension packages using LessCSS or UglifyJS.
Extension package building hit a regression when building against
installed versions of a product utilizing Djblets. Pipeline was looking
for LessCSS and UglifyJS relative to the product's node_modules
directory, rather than the one generated for the extension.

The reason this was happening was that the Pipeline settings being set
were for Pipeline 1.6.x, which is used on Djblets 0.10, and not Pipeline
1.3.x, which is used on Djblets 0.9. This meant that the defaults for
the product were being used instead.

Those defaults could also be overridden. We had code designed to
re-populate Pipeline's versions of the settings based on what was in the
product's settings, but these were overriding the newly-set values. We
now re-populate before computing new values.

The latter part applies to Djblets 0.9 and higher. The settings names
are specific to Djblets 0.9 and will not be merged up to 0.10.

Testing Done:
Reproduced the failure with an extension. Verified it was fixed with
this patch.

Reviewed at https://reviews.reviewboard.org/r/9014/