Add whitespace stripping to the {% attr %} template tag.
The `{% attr %}` template tag now strips whitespace at the beginning and
end of the value. It used to do this when considering the contents, but
then preserved the whitespace when injecting it into the value. Since
it's unlikely callers (which tend to use this for CSS class names) want
to retain leading/trailing whitespace for rendered values, the
whitespace is no longer preserved.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9051/