Support rendering with additional context in Config.ListItemView
`Djblets.Config.ListItemView` now supports an additional method,
`getRenderContext` that will be used with the associated model's
attributes to generate the full render context for the template. This
allows subclasses to inject things into the template that are not
necessarily model attributes (but my be derived from them).

Testing Done:
Tested this with an upcoming patch for Review Board.

Reviewed at https://reviews.reviewboard.org/r/9075/