Refactor config forms LESS to use variables, support empty item lists
This patch refactors the config forms LESS to use defined variables for
colours. These colours were being used in Review Board's account pages
and this change will allow the colours to be imported (instead of the
current hardcoding that occurs).

The CSS for config forms lists has been slightly tweaked so that they
have the same top and bottom border. Previously, there was no bottom
border on the list -- it was just the `border-bottom` from the last
element in the list, which uses a different colour from the `border-top`
of the entire list.

Additionally, support has been added for a new
`.config-forms-list-empty` element that is intended to be a sibling of
the `.config-forms-list`. It will only be shown when the list has n
children. This is a generally useful pattern for supporting
placeholders that don't have to be dynamically added and removed from
the page. This is a refactor from a pattern I was using in Review
Board.

Testing Done:
Tested with an upcoming change for Review Board.

Reviewed at https://reviews.reviewboard.org/r/9074/