Downgrade ResourceOAuth2TokenMixin error to warning
The OAuth2 scopes code logs an error if a resource does not inherit fom
the OAuth2 token mixin. We are downgrading this to a warning to seem
less scary (and because its not actually an error).

Testing Done:
Confirmed the error no longer appears and a warning appears instead.

Reviewed at https://reviews.reviewboard.org/r/9124/