Always call parent setUp/tearDown methods in test suites.
We had a number of test suites that weren't calling their parent class's
`setUp()` or `tearDown()` methods. Strictly speaking, it's not necessary
to call these in these cases, but it will be important down the road if
there are any changes in the inheritance. To keep things consistent and
safe going forward, the parent methods are now called in every suite.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9230/