Add support for enabling certain extensions by default.
This adds a new setting, `settings.EXTENSIONS_ENABLED_BY_DEFAULT`, which
is a list of extension IDs that will be automatically enabled when first
registered in the database through `ExtensionManager.load()`. This
allows a project to automatically enable specific extensions, such as
first-party ones considered a part of the application.

Once an extension is later disabled, it will not be automatically
re-enabled unless the registration is removed from the database.
Extensions also cannot opt into being enabled by default.

Testing Done:
Unit tests pass.

Tested this manually with an extension that was set to be auto-enabled.

Reviewed at https://reviews.reviewboard.org/r/9273/