Add an "id" property for integration instances.
This adds an `id` property on integrations that works as an alias for
the `integration_id` property. While the latter is preferred, this alias
is necessary for some hooks to retain compatibiliy with both extensions
(which use `id`) and integrations (which did not).

Testing Done:
Made use of this with an integration and a hook that needed the ID. It
didn't crash.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9284/