Clean up dead space and clipped icons in the datagrid's Edit Columns menu.
The styling for the Edit Columns menu for datagrids resulted in a lot of
unclickable dead space around icons and text, requiring more precise
clicking in order to add or remove columns. Along with this, the
checkmark icons weren't properly aligned and were actually clipped a
bit, based on their position in the spritesheet.

This change tightens up the spacing, removing all the dead space so any
click anywhere on a row will add or remove the column. It also removes
unwanted whitespace before the labels, fixes the referenced position of
the checkmark icon in the spritesheet, and fixes alignment of the icon.

To improve accessibility and to better inform users just a bit, a title
has been given to the row indicating if a click will add or remove the
column.

Testing Done:
Checked that there was no longer any dead space within a row.

Checked the title on the cells, icons, and text.

Checked the alignment of the checkmarks and the in-label icons to make
sure things were consistent.

Tested clicking on the checkmark area and the label to make sure
columns were added and removed.

Reviewed at https://reviews.reviewboard.org/r/9300/