Fix a regression with an attribute in SettingsForm for extensions.
Extensions used to have access to a `siteconfig` attribute on
`SettingsForm`, which was were the settings for the extension lived. The
name was there for historical reasons, due to the form originally being
made for `SiteConfiguration`. The move to using the new `KeyValueForm`
in Djblets 1.0 resulted in this attribute being removed, breaking
extensions.

This change restores the attribute, emitting a `DeprecationWarning` when
accessed. The warning is only emitted once per access per form
instance, in order to avoid spamming the logs.

Testing Done:
Tested an extension with Review Board that made use of this. Verified that
it no longer broke.

Reviewed at https://reviews.reviewboard.org/r/9391/