Add release notes for Djblets 1.0.1.
This is a small release focusing only on the fix for the disappearance
of avatar backends. Other fixes that are in progress will be pushed to
future patch releases.

Testing Done:
Build the docs and read through them. Checked for spelling and build
errors.

Reviewed at https://reviews.reviewboard.org/r/9422/