Add a line-height to .avatar-missing.
The missing avatar display didn't define the line height, which meant we
were at the mercy of the element's inherted vertical alignment. This
change fixes that so that the size of the question mark element is
always the same height as its container.

Testing done:
Saw the correct alignment in both Review Board and elsewhere.

Reviewed at https://reviews.reviewboard.org/r/9465/