Prevent gravatars from exploding if users don't have an email address.
The recent changes to the gravatar backend were causing uncaught
ValueErrors to percolate upwards for users with empty email addresses,
breaking a variety of things. Not having an avatar shouldn't be a fatal
condition, so I've changed it to return an empty dict instead.

Testing done:
- Ran unit tests.
- Tested in conjunction with some other stuff that was previously
  breaking.

Reviewed at https://reviews.reviewboard.org/r/9491/